refactor: SchemaType should be a SchemaItem
- `type_api.TypeEngineMixin` should be `visitors.Visitable` to
  ensure `Visitable` always appears last when linearizing
  `TypeEngine` subclasses.
  (Actually, Visitable should arguably be a metacls, but I'm
   not confident to make that kind of change)
- `metadata` is always set on `SchemaType` when a value is
  available
- Add internal `SchemaType._set_metadata` method to clean up
  event handling
